+++ src/DependencyInjection/Configuration.php
@@ -6,6 +6,7 @@ namespace Dbp\Relay\AuthorizationBundle\DependencyInjection;
 
 use Dbp\Relay\AuthorizationBundle\Authorization\AuthorizationService;
 use Dbp\Relay\AuthorizationBundle\Entity\GrantedActions;
+use Dbp\Relay\AuthorizationBundle\Service\UserAttributeProvider;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -79,7 +80,13 @@ class Configuration implements ConfigurationInterface
                                 ->ifTrue(function ($value) {
                                     return str_contains($value, GrantedActions::ID_SEPARATOR);
                                 })
-                                ->thenInvalid('Resource class identifiers must not contain reserved separator character \''.GrantedActions::ID_SEPARATOR.'\'')
+                                ->thenInvalid('Resource class identifiers must not contain reserved character \''.GrantedActions::ID_SEPARATOR.'\'')
+                                ->end()
+                                ->validate()
+                                ->ifTrue(function ($value) {
+                                    return str_contains($value, UserAttributeProvider::SEPARATOR);
+                                })
+                                ->thenInvalid('Resource class identifiers must not contain reserved character \''.UserAttributeProvider::SEPARATOR.'\'')
                                 ->end()
                                 ->info('The resource class identifier')
                             ->end()